feat: rename trainer_builder to builders
